servers/estate-planning-mcp/pyproject.toml
[project]
name = "estate-planning-mcp"
version = "0.1.0"
description = "Estate-planning MCP server for succession, trusts, entities, ownership, and legal docs"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "jsonschema>=4.23.0",
    "stewardos-lib",
]

[tool.uv.sources]
stewardos-lib = { path = "../lib", editable = true }

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
